FileShare/requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.3
django-session-timeout==0.1.0
djangorestframework==3.14.0
filelock==3.8.0
multipledispatch==0.6.0
mysql-connector==2.2.9
mysql-connector-python==8.0.30
Pillow==9.3.0
platformdirs==2.5.4
playsound==1.3.0
prettytable==3.4.1
protobuf==3.20.1
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.16.7
wcwidth==0.2.5
